RegressionLearnersTest: Remove pointless test, Refactor
